feat: add more testing
