+++ src/utils/test-utils/index.tsx
@@ -1,3 +1,7 @@
 export { mountWithContext, mountWithDrawerSuspense } from "./mountWith";
 export { mockAlbumDiaryWithId, mockAlbumFetchData } from "./dataSpotify";
-export { mockMovieDiaryWithId, mockMovieFetchData } from "./dataMovie";
+export {
+  mockMovieDiaryWithId,
+  mockMovieFetchData,
+  mockMovieSelected,
+} from "./dataMovie";
